Move src/pygame to display
 - and add __init__.py - files
